pyproject.toml
[project]
name = "hexlet-code"
version = "0.1.0"
description = "Task Manager Django project"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dj-database-url>=3.0.1",
    "django>=5.2.6",
    "django-bootstrap5>=25.2",
    "django-filter>=25.2",
    "gunicorn>=23.0.0",
    "python-dotenv>=1.1.1",
    "whitenoise>=6.11.0",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "pytest-django>=4.11.1",
    "ruff>=0.14.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["task_manager"]

[tool.pytest.ini_options]
testpaths = ["task_manager"]
python_files = ["tests.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
DJANGO_SETTINGS_MODULE = "task_manager.settings"
